+++ src/test/java/benchmark/CreateSchemaBenchmark.java
@@ -16,34 +16,24 @@ import org.openjdk.jmh.infra.Blackhole;
 
 import java.util.concurrent.TimeUnit;
 
-/**
- * This benchmarks schema creation
- * <p>
- * See https://github.com/openjdk/jmh/tree/master/jmh-samples/src/main/java/org/openjdk/jmh/samples/ for more samples
- * on what you can do with JMH
- * <p>
- * You MUST have the JMH plugin for IDEA in place for this to work :  https://github.com/artyushov/idea-jmh-plugin
- * <p>
- * Install it and then just hit "Run" on a certain benchmark method
- */
 @Warmup(iterations = 2, time = 5)
 @Measurement(iterations = 3)
 @Fork(3)
-public class SchemaBenchMark {
+public class CreateSchemaBenchmark {
 
     static String largeSDL = BenchmarkUtils.loadResource("large-schema-3.graphqls");
 
     @Benchmark
     @BenchmarkMode(Mode.Throughput)
     @OutputTimeUnit(TimeUnit.MINUTES)
-    public void benchMarkLargeSchemaCreate(Blackhole blackhole) {
+    public void benchmarkLargeSchemaCreate(Blackhole blackhole) {
         blackhole.consume(createSchema(largeSDL));
     }
 
     @Benchmark
     @BenchmarkMode(Mode.AverageTime)
     @OutputTimeUnit(TimeUnit.MILLISECONDS)
-    public void benchMarkLargeSchemaCreateAvgTime(Blackhole blackhole) {
+    public void benchmarkLargeSchemaCreateAvgTime(Blackhole blackhole) {
         blackhole.consume(createSchema(largeSDL));
     }
