+++ src/test/java/benchmark/DFSelectionSetBenchmark.java
@@ -21,7 +21,6 @@ import org.openjdk.jmh.annotations.OutputTimeUnit;
 import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
-import org.openjdk.jmh.annotations.Threads;
 import org.openjdk.jmh.annotations.Warmup;
 import org.openjdk.jmh.infra.Blackhole;
 
@@ -65,10 +64,6 @@ public class DFSelectionSetBenchmark {
     }
 
     @Benchmark
-    @Warmup(iterations = 2)
-    @Measurement(iterations = 5, time = 10)
-    @Threads(1)
-    @Fork(3)
     @BenchmarkMode(Mode.AverageTime)
     @OutputTimeUnit(TimeUnit.MILLISECONDS)
     public void benchMarkAvgTime(MyState myState, Blackhole blackhole) {
@@ -77,10 +72,6 @@ public class DFSelectionSetBenchmark {
     }
 
     @Benchmark
-    @Warmup(iterations = 2)
-    @Measurement(iterations = 5, time = 10)
-    @Threads(1)
-    @Fork(3)
     @BenchmarkMode(Mode.Throughput)
     @OutputTimeUnit(TimeUnit.MILLISECONDS)
     public void benchMarkThroughput(MyState myState, Blackhole blackhole) {
